Refactor filter rule parsing module
