+++ crates/cli/src/frontend/filter_rules/parsing/shorthand.rs
@@ -1,0 +1,43 @@
+use oc_rsync_core::client::FilterRuleSpec;
+use oc_rsync_core::message::{Message, Role};
+use oc_rsync_core::rsync_error;
+
+use super::super::directive::FilterDirective;
+use super::helpers::trim_short_rule_remainder;
+
+pub(super) fn parse_filter_shorthand(
+    trimmed: &str,
+    short: char,
+    label: &str,
+    builder: fn(String) -> FilterRuleSpec,
+) -> Option<Result<FilterDirective, Message>> {
+    let mut chars = trimmed.chars();
+    let first = chars.next()?;
+    if !first.eq_ignore_ascii_case(&short) {
+        return None;
+    }
+
+    let remainder = chars.as_str();
+    if remainder.is_empty() {
+        let text = format!("filter rule '{trimmed}' is missing a pattern after '{label}'");
+        let message = rsync_error!(1, text).with_role(Role::Client);
+        return Some(Err(message));
+    }
+
+    if !remainder
+        .chars()
+        .next()
+        .is_some_and(|ch| ch.is_ascii_whitespace() || ch == '_')
+    {
+        return None;
+    }
+
+    let pattern = trim_short_rule_remainder(remainder);
+    if pattern.is_empty() {
+        let text = format!("filter rule '{trimmed}' is missing a pattern after '{label}'");
+        let message = rsync_error!(1, text).with_role(Role::Client);
+        return Some(Err(message));
+    }
+
+    Some(Ok(FilterDirective::Rule(builder(pattern.to_string()))))
+}
